naturedrinkbackend/requirement.txt
Django==1.10.2
django-cors-headers==1.2.1
django-rest==0.0.1
djangorestframework==3.4.7
mysqlclient==1.3.7
